fix reset animation in hero section
